Bug: add debug

+++ pg_salience_feature/inference/constrained_beam_search.py
@@ -318,9 +318,9 @@ class ConstrainedBeamSearch:
         best_word_indices_list = []
 
         lengths = torch.zeros((batch_size*self.beam_size, 1), device=device)
-        finished = torch.zeros((batch_size*self.beam_size, ), device=device)
+        finished = torch.zeros((batch_size*self.beam_size, ), device=device, dtype=torch.int32)
         is_pad_dist_set = False
-        scores_accumulated = torch.zeros((batch_size * self.beam_size, 1), dtype=torch.int32,device=device)
+        scores_accumulated = torch.zeros((batch_size * self.beam_size, 1), device=device)
         inactive = torch.zeros((batch_size * self.beam_size), dtype=torch.int32, device=device)
         for t in range(1, self.max_steps):
             # shape: (batch_size, num_classes)
